support statistics for iceberg, paimon table
